+++ src/components/pages/reviews-page.tsx
@@ -1,15 +1,15 @@
 'use client'
 
 import Link from 'next/link'
-import { useReviews } from '@/hooks/useReviews'
+import { useReviews } from '@/app/(menu)/pools/[id]/reviews/_hooks/useReviews'
 import { Separator } from '@/components/ui/separator'
-import ReviewItem from '@/components/reviews/review-item'
+import ReviewItem from '@/app/(menu)/pools/[id]/reviews/_components/review-item'
 import { ResponsiveDialog } from '@/components/responsive-dialog'
 import { useState } from 'react'
-import ReviewForm from '@/components/reviews/reivew-post-form'
+import ReviewForm from '@/app/(menu)/pools/[id]/reviews/_components/reivew-post-form'
 import { Button } from '@/components/ui/button'
 import { PiPencilSimpleLineFill } from 'react-icons/pi'
-import { IReview } from '@/types/reviews'
+import { IReview } from '@/app/(menu)/pools/[id]/reviews/_types/reviews'
 import { FaChevronLeft } from 'react-icons/fa'
 import { useIntersectionObserver } from '@/hooks/use-intersectionObserver'
